refactor about gig component into two
